requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.9
django-adminlte-ui==1.5.1
django-autocomplete-light==3.8.2
django-crispy-forms==1.14.0
django-import-export==2.7.1
django-mathfilters==1.0.0
django-treebeard==4.3
django-widget-tweaks==1.4.11
et-xmlfile==1.1.0
future==0.18.2
html5lib==1.1
idna==3.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.0
psycopg2==2.9.3
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.3
PyYAML==6.0
reportlab==3.6.6
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
tzdata==2021.5
urllib3==1.26.8
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
